Squashed commit of the following:
commit 56cc21de975fb26fe6abc573faceb015326b133d
Author: gamberfox <gamberfox@gmail.com>
Date:   Tue Aug 20 16:15:45 2024 -0500

    Squashed commit of the following:

    commit 74f9c9aafdfba724279804c724657779ecb3d276
    Author: gamberfox <gamberfox@gmail.com>
    Date:   Tue Aug 20 16:14:56 2024 -0500

        HU-003 progress

commit 542f98bf135965eea47dcf507ef902f0a0fbc42a
Author: gamberfox <gamberfox@gmail.com>
Date:   Tue Aug 20 15:04:23 2024 -0500

    Squashed commit of the following:

    commit e0109cd17a0666228cc238f031eb1fff841a5e59
    Author: gamberfox <gamberfox@gmail.com>
    Date:   Tue Aug 20 15:03:01 2024 -0500

        HU-002 progress

    commit 6a3681a406082fbe8baf79ab242198b0269c8ee5
    Author: gamberfox <gamberfox@gmail.com>
    Date:   Tue Aug 20 14:25:53 2024 -0500

        HU-002 progress
        implemented pagination

    commit d4b6432f1c28a61b324c9aaedb47b2d55483e09f
    Author: gamberfox <gamberfox@gmail.com>
    Date:   Tue Aug 20 13:58:02 2024 -0500

        HU-002 progress

        I'm trying to set up pagination

    commit bd06eecfecec8f746e6443065e7da39b58f22636
    Author: gamberfox <gamberfox@gmail.com>
    Date:   Tue Aug 20 10:11:01 2024 -0500

        HU-002 progress

        I'm trying to set up pagination

    commit 6b1ab97ce95a079fa419440b1900575fda9a7f70
    Author: gamberfox <gamberfox@gmail.com>
    Date:   Tue Aug 20 06:16:50 2024 -0500

        merging dev into HU-002 branch

commit 70caf02a8b3180f31a0c6a29ec9220cd28fbbc83
Author: gamberfox <gamberfox@gmail.com>
Date:   Tue Aug 20 06:15:41 2024 -0500

    merging HU-001 into dev branch

commit 9554c8f0b202ec710d4eddaad4d16e26df989c06
Merge: 0aac72b 0c72e6e
Author: gamberfox <gamberfox@gmail.com>
Date:   Mon Aug 19 13:39:06 2024 -0500

    Merge branch 'dev' of github.com:gamberfox/pragma-bootcamp-project into dev

commit 0c72e6e6a50cc616de8dbf253183d17daeb38293
Merge: 77aa795 4ab0dce
Author: gamberfox <gamberfox@gmail.com>
Date:   Sun Aug 18 21:12:30 2024 -0500

    merging HU-001 into main

commit 4ab0dcef7b0b3ed961623353f441fbc67a601a1d
Author: gamberfox <gamberfox@gmail.com>
Date:   Sun Aug 18 21:04:46 2024 -0500

    main branch update

+++ stock.sql
@@ -8,32 +8,32 @@ USE stock;
 -- brand table
 DROP TABLE IF EXISTS brand;
 CREATE TABLE brand (
-    id INT AUTO_INCREMENT PRIMARY KEY,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(50) UNIQUE,
     description VARCHAR(120)
 );
 -- category table
 DROP TABLE IF EXISTS category;
 CREATE TABLE category (
-    id INT AUTO_INCREMENT PRIMARY KEY,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(50) unique,
     description VARCHAR(90) NOT NULL
 );
 -- article table
 DROP TABLE IF EXISTS article;
 CREATE TABLE article (
-    id INT AUTO_INCREMENT PRIMARY KEY,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(50) unique,
-    stock_quantity INT,
+    stock_quantity BIGINT,
     price decimal(10, 2),
-    id_brand INT,
+    id_brand BIGINT,
     FOREIGN KEY (id_brand) REFERENCES brand(id)
 );
 -- article_category table
 DROP TABLE IF EXISTS article_category;
 CREATE TABLE article_category (
-    id_article INT NOT NULL,
-    id_category INT NOT NULL,
+    id_article BIGINT NOT NULL,
+    id_category BIGINT NOT NULL,
     PRIMARY KEY (id_article, id_category),
     FOREIGN KEY (id_article) REFERENCES article(id),
     FOREIGN KEY (id_category) REFERENCES category(id)
